Moved truncateCrumbTrail from StateHandler to State
