+++ NavigationJS/test/NavigationTest.ts
@@ -1470,7 +1470,7 @@ describe('Navigation', function () {
                     { key: 's', route: 'r', trackCrumbTrail: true },
                 ]);
             var state = stateNavigator.states['s'];
-            state.stateHandler.truncateCrumbTrail = (state, crumbs) => {
+            state.truncateCrumbTrail = (state, crumbs) => {
                 return crumbs;
             };
         });
@@ -1509,7 +1509,7 @@ describe('Navigation', function () {
                 { key: 's1', route: 'r1', trackCrumbTrail: true }
             ]);
             var state = stateNavigator.states['s1'];
-            state.stateHandler.truncateCrumbTrail = (state, crumbs) => {
+            state.truncateCrumbTrail = (state, crumbs) => {
                 return crumbs;
             };
         });
@@ -1551,7 +1551,7 @@ describe('Navigation', function () {
                 { key: 's', route: 'r', trackCrumbTrail: true }
             ]);
             var state = stateNavigator.states['s'];
-            state.stateHandler.truncateCrumbTrail = (state, crumbs) => {
+            state.truncateCrumbTrail = (state, crumbs) => {
                 return crumbs;
             };
         });
@@ -1595,7 +1595,7 @@ describe('Navigation', function () {
                 { key: 's1', route: 'r1', trackCrumbTrail: true }
             ]);
             var state = stateNavigator.states['s1'];
-            state.stateHandler.truncateCrumbTrail = (state, crumbs) => {
+            state.truncateCrumbTrail = (state, crumbs) => {
                 return crumbs;
             };
         });
@@ -1643,7 +1643,7 @@ describe('Navigation', function () {
                 { key: 's2', route: 'r2', trackCrumbTrail: true }
             ]);
             var state = stateNavigator.states['s1'];
-            state.stateHandler.truncateCrumbTrail = (state, crumbs) => {
+            state.truncateCrumbTrail = (state, crumbs) => {
                 return crumbs;
             };
         });
@@ -1694,7 +1694,7 @@ describe('Navigation', function () {
                 { key: 's2', route: 'r2', trackCrumbTrail: true }
             ]);
             var state = stateNavigator.states['s1'];
-            state.stateHandler.truncateCrumbTrail = (state, crumbs) => {
+            state.truncateCrumbTrail = (state, crumbs) => {
                 return crumbs;
             };
         });
@@ -4424,7 +4424,7 @@ describe('Navigation', function () {
                 { key: 's1', route: 'r1', trackCrumbTrail: true }
             ]);
             var state = stateNavigator.states['s1'];
-            state.stateHandler.truncateCrumbTrail = (state, crumbs) => {
+            state.truncateCrumbTrail = (state, crumbs) => {
                 return crumbs;
             };
         });
